+++ lib/country.ts
@@ -3,8 +3,8 @@ import { findEntryByCode } from '../utils';
 import { ICountry } from './interface';
 
 // Get a country by isoCode.
-function getCountryByCode(isoCode: string): ICountry | '' {
-	if (!isoCode) return '';
+function getCountryByCode(isoCode: string): ICountry | undefined {
+	if (!isoCode) return undefined;
 
 	return findEntryByCode(countryList, isoCode);
 }