+++ lib/state.ts
@@ -9,6 +9,7 @@ export function getAllStates(): IState[] {
 
 // Get a list of states belonging to a specific country.
 export function getStatesOfCountry(countryCode: string): IState[] {
+	if (!countryCode) return [];
 	const states = stateList.filter((value) => {
 		return value.countryCode === countryCode;
 	});
@@ -16,20 +17,20 @@ export function getStatesOfCountry(countryCode: string): IState[] {
 }
 
 // Find a country by it's ISO code and the country in which it is contained.
-export function getStateByCodeAndCountry(stateCode: string, countryCode: string): IState | '' {
-	if (!stateCode) return '';
-	if (!countryCode) return '';
+export function getStateByCodeAndCountry(stateCode: string, countryCode: string): IState | undefined {
+	if (!stateCode) return undefined;
+	if (!countryCode) return undefined;
 
 	return findStateByCodeAndCountryCode(stateList, stateCode, countryCode);
 }
 
 // to be deprecate
-export function getStateByCode(isoCode: string): IState | '' {
+export function getStateByCode(isoCode: string): IState | undefined {
 	// eslint-disable-next-line no-console
 	console.warn(
 		`WARNING! 'getStateByCode' has been deprecated, please use the new 'getStateByCodeAndCountry' function instead!`,
 	);
-	if (!isoCode) return '';
+	if (!isoCode) return undefined;
 
 	return findEntryByCode(stateList, isoCode);
 }