+++ utils/index.ts
@@ -3,9 +3,9 @@ export const findEntryByCode = (source: any, code: string) => {
 		const codex = source.findIndex((c: any) => {
 			return c.isoCode === code;
 		});
-		return codex !== -1 ? source[codex] : '';
+		return codex !== -1 ? source[codex] : undefined;
 	}
-	return '';
+	return undefined;
 };
 
 export const findStateByCodeAndCountryCode = (source: any, code: string, countryCode: string) => {
@@ -13,9 +13,9 @@ export const findStateByCodeAndCountryCode = (source: any, code: string, country
 		const codex = source.findIndex((c: any) => {
 			return c.isoCode === code && c.countryCode === countryCode;
 		});
-		return codex !== -1 ? source[codex] : '';
+		return codex !== -1 ? source[codex] : undefined;
 	}
-	return '';
+	return undefined;
 };
 
 export const compare = (a: any, b: any) => {